#160 inital commit of Romans mecab code
plugin-mecab-preprocessing is made a
standalone project. It will not compile together
with the module project TopicExplorer. Instead
build it separately with maven. It does build fine.
However the implemented test will fail as it requires
a mecab 0.994 installation. Some paths in the code
will need to be adapted to find the mecab installation.