TreeTagger is integrated as part of plugin-mecab-initcorpus. It uses the lucene-analyzer-treetagger by Alexander Hinneburg that transforms the output of treetagger into lucence analyzer token stream. This would prove helpful, when lucene is used for keyword searching instead of mysql.
At the moment, there is a switch in mecab.global.properties that allows
to choose among mecab and treetagger with German and English dictionary.
For those two dicts the respective table contents of PosType are
provided in PosTypeFill command.

The table PosType is extended by attribute LONG_DESCRIPTION that
contains a verbal explanation of the POS tag. This will be used by
webapp-preprocessing for explaining POS tags.

The current state allows to run webapp-prepocessing without
modification. A refactoring of plugin-mecab-initcorpus is neccessary to
isolate common parts of text preprocessing and the mecab and treetagger
specifics.