#239 plumbing DocumentChunk implementation into commands
The new table commands DocumentChunkCreate and DocumentChunkFill
integrate the implementation for computing and storing document chunks
from class DocumentChunk into the commandManager workflow.

Tests are added to the implementation to cover wrong behavior that is
now fixed in class DocumentChunk.

Reading delimiters from XML file was wrong in the original place
DelimiterPositionsFill. Changes to correct this:
- increase Apache.commons.config to version 1.10 in pom
- Change reading mode setDelimiterParsingDisabled to true. Otherwise
comma delimiter is gobbled by Apache.common.config, which is the default
behavior, (furhter details, see documention there).
- use getStringArray instead of getList (this had to be changed in
DelimiterPositionsFill as well to ensure it still compiles, however,
this class will be deleted in next refactoring, thus, now further
changes are made)

Used delimiters are analyzed and a message is written to the log file,
which is helpful to report configurated delimiters by users.

- command catalogs preJooqConfig.xml and postJooqConfig.xml are changed
to excute only the DocumentChunk functionality.